Added exception handling

+++ src/main/java/com/example/csmaster/service/CustomerService.java
@@ -1,6 +1,7 @@
 package com.example.csmaster.service;
 
 import com.example.csmaster.model.Customer;
+import com.example.csmaster.exception.CustomerNotFoundException;
 import com.example.csmaster.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,7 @@ public class CustomerService {
     private CustomerRepository customerRepository;
 
     public Customer findCustomerById(Long id) {
+        if(!customerRepository.existsById(id)) throw new CustomerNotFoundException();
         return customerRepository.findById(id).get();
     }
     public List<Customer> findAll() {
@@ -28,6 +30,8 @@ public class CustomerService {
     }
 
     public Customer updateCustomer(Long id, Customer newDetails) {
+        if(!customerRepository.existsById(id)) throw new CustomerNotFoundException();
+
         Customer customer = customerRepository.findById(id).get();
         customer.setName(newDetails.getName());
         customer.setDescriptions(newDetails.getDescriptions());
@@ -36,6 +40,7 @@ public class CustomerService {
     }
 
     public String deleteCustomerById(Long id) {
+        if(!customerRepository.existsById(id)) throw new CustomerNotFoundException();
         customerRepository.deleteById(id);
         return "User with id " + id + " is deleted.";
     }
